refactor tests into separate files
